Renamed types.py to sbi_types.py
