+++ sbi/inference/posteriors/direct_posterior.py
@@ -12,8 +12,8 @@ from sbi.inference.potentials.posterior_based_potential import (
 )
 from sbi.neural_nets.density_estimators.base import DensityEstimator
 from sbi.samplers.rejection.rejection import accept_reject_sample
-from sbi.types import Shape
-from sbi.utils import check_prior, within_support
+from sbi.sbi_types import Shape
+from sbi.utils import check_prior, match_theta_and_x_batch_shapes, within_support
 from sbi.utils.torchutils import ensure_theta_batched
