+++ sbi/inference/potentials/ratio_based_potential.py
@@ -8,7 +8,7 @@ from torch import Tensor, nn
 from torch.distributions import Distribution
 
 from sbi.inference.potentials.base_potential import BasePotential
-from sbi.types import TorchTransform
+from sbi.sbi_types import TorchTransform
 from sbi.utils import mcmc_transform
 from sbi.utils.sbiutils import match_theta_and_x_batch_shapes
 from sbi.utils.torchutils import atleast_2d
