+++ sbi/neural_nets/density_estimators/nflows_flow.py
@@ -5,7 +5,7 @@ from pyknos.nflows.flows import Flow
 from torch import Tensor, nn
 
 from sbi.neural_nets.density_estimators.base import DensityEstimator
-from sbi.types import Shape
+from sbi.sbi_types import Shape
 
 
 class NFlowsFlow(DensityEstimator):
