+++ tests/test_utils.py
@@ -13,7 +13,7 @@ from sbi.inference.posteriors.base_posterior import NeuralPosterior
 from sbi.inference.posteriors.direct_posterior import DirectPosterior
 from sbi.inference.potentials.base_potential import BasePotential
 from sbi.simulators.linear_gaussian import true_posterior_linear_gaussian_mvn_prior
-from sbi.types import Shape, TorchTransform
+from sbi.sbi_types import Shape, TorchTransform
 from sbi.utils import BoxUniform, within_support
 from sbi.utils.metrics import c2st
 from sbi.utils.torchutils import ensure_theta_batched
